test: move sed actor replay to nightly
