+++ packages/agentos/package.json
@@ -51,7 +51,8 @@
 		"build:actor": "tsup src/index.ts src/client.ts src/react.ts",
 		"build:tabs": "vite build --config vite.tabs.config.ts",
 		"check-types": "tsc --noEmit",
-		"test": "vitest run",
+		"test": "vitest run --exclude '**/*.nightly.test.ts'",
+		"test:nightly": "AGENTOS_ACTOR_E2E=1 vitest run tests/*.nightly.test.ts --fileParallelism=false --passWithNoTests",
 		"test:e2e": "cargo build --manifest-path ../../Cargo.toml -p agentos-sidecar && pnpm --filter @rivet-dev/agentos-runtime-core build && pnpm --filter @agentos-software/manifest build && pnpm --filter @agentos-software/common... build && pnpm --filter @agentos-software/coreutils build:runtime && pnpm --filter @rivet-dev/agentos-test-harness build && pnpm --filter @rivet-dev/agentos-core build && pnpm build && pnpm test:e2e:run",
 		"test:e2e:run": "AGENTOS_ACTOR_E2E=1 vitest run tests/actor.e2e.test.ts tests/agent-os-conformance.e2e.test.ts --fileParallelism=false"
 	},
